# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Short.ly API Documentation",
        "version": "1.0.0",
        "description": "API documentation for the Short.ly URL shortener service",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "contact": {
            "name": "API Support",
            "url": "https://short.ly/support",
            "email": "support@short.ly"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development server"
        },
        {
            "url": "https://api-mvp.short.ly",
            "description": "MVP server"
        },
        {
            "url": "https://api.short.ly",
            "description": "Production server"
        }
    ],
...
    "paths": {
        "/api/v1/urls": {
            "post": {
                "summary": "Create a new short URL",
                "tags": [
                    "URLs"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "originalUrl"
                                ],
                                "properties": {
                                    "originalUrl": {
                                        "type": "string",
                                        "description": "The original URL to shorten",
                                        "example": "https://example.com/very/long/path/to/page.html"
                                    },
                                    "customAlias": {
                                        "type": "string",
                                        "description": "Optional custom alias",
                                        "example": "my-page"
                                    },
                                    "expiresAt": {
                                        "type": "string",
                                        "format": "date-time",
                                        "description": "Optional expiration date",
                                        "example": "2023-12-31T23:59:59Z"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "URL created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Url"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Custom alias already in use",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Get all URLs (admin only)",
                "tags": [
                    "URLs"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        },
                        "description": "Page number"
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "minimum": 1,
                            "maximum": 100
                        },
                        "description": "Results per page"
                    },
                    {
                        "in": "query",
                        "name": "isActive",
                        "schema": {
                            "type": "boolean"
                        },
                        "description": "Filter by active status"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "URLs retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "urls": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Url"
                                                    }
                                                },
                                                "pagination": {
                                                    "type": "object",
                                                    "properties": {
                                                        "totalCount": {
                                                            "type": "integer",
                                                            "example": 42
                                                        },
                                                        "page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "totalPages": {
                                                            "type": "integer",
                                                            "example": 5
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Admin access required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/urls/{shortCode}": {
            "get": {
                "summary": "Get details of a URL by short code",
                "tags": [
                    "URLs"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "shortCode",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The short code of the URL"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "URL details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Url"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "URL not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "summary": "Update a URL",
                "tags": [
                    "URLs"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "shortCode",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The short code of the URL"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "isActive": {
                                        "type": "boolean",
                                        "description": "Set the URL active status"
                                    },
                                    "expiresAt": {
                                        "type": "string",
                                        "format": "date-time",
                                        "description": "Update expiration date"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "URL updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Url"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "URL not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permission denied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete a URL",
                "tags": [
                    "URLs"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "shortCode",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The short code of the URL"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "URL deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "URL deleted successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "URL not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permission denied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/urls/user/me": {
            "get": {
                "summary": "Get URLs for the authenticated user",
                "tags": [
                    "URLs"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        },
                        "description": "Page number"
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "minimum": 1,
                            "maximum": 100
                        },
                        "description": "Results per page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "URLs retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "urls": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Url"
                                                    }
                                                },
                                                "pagination": {
                                                    "type": "object",
                                                    "properties": {
                                                        "totalCount": {
                                                            "type": "integer",
                                                            "example": 42
                                                        },
                                                        "page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "totalPages": {
                                                            "type": "integer",
                                                            "example": 5
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/urls/alias/{alias}/check": {
            "get": {
                "summary": "Check if a custom alias is available",
                "tags": [
                    "URLs"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "alias",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The custom alias to check"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alias availability checked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "alias": {
                                                    "type": "string",
                                                    "example": "my-page"
                                                },
                                                "isAvailable": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Url": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "60d21b4667d0d8992e610c85"
                    },
                    "originalUrl": {
                        "type": "string",
                        "example": "https://example.com/very/long/path/to/page.html"
                    },
                    "shortCode": {
                        "type": "string",
                        "example": "2g3ab7c"
                    },
                    "shortUrl": {
                        "type": "string",
                        "example": "https://short.ly/2g3ab7c"
                    },
                    "customAlias": {
                        "type": "string",
                        "example": "my-page",
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "expiresAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "isActive": {
                        "type": "boolean",
                        "example": true
                    },
                    "metadata": {
                        "type": "object",
                        "properties": {
                            "totalClicks": {
                                "type": "integer",
                                "example": 42
                            },
                            "lastAccessed": {
                                "type": "string",
                                "format": "date-time",
                                "nullable": true
                            }
                        }
                    }
                }
            },
...
            "Error": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "object",
                        "properties": {
                            "message": {
                                "type": "string",
                                "example": "An error occurred"
                            },
                            "status": {
                                "type": "integer",
                                "example": 400
                            },
                            "errors": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "param": {
                                            "type": "string",
                                            "example": "originalUrl"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Invalid URL format"
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "body"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}